Move event stuff in service.event
