+++ MANO-API/src/main/java/com/ubiqube/etsi/mano/controller/nslcm/sol003/VnfLcmSol003Api.java
@@ -26,8 +26,8 @@ import com.ubiqube.etsi.mano.model.nslcm.sol003.TerminateVnfRequest;
 import com.ubiqube.etsi.mano.model.nslcm.sol003.VnfInstance;
 import com.ubiqube.etsi.mano.model.nslcm.sol003.VnfInstance.InstantiationStateEnum;
 import com.ubiqube.etsi.mano.repository.VnfInstancesRepository;
-import com.ubiqube.etsi.mano.service.EventManager;
-import com.ubiqube.etsi.mano.service.NotificationEvent;
+import com.ubiqube.etsi.mano.service.event.EventManager;
+import com.ubiqube.etsi.mano.service.event.NotificationEvent;
 
 @Profile({ "default", "VNFM" })
 @RestController
