+++ MANO-API/src/test/java/com/ubiqube/etsi/mano/controller/vnf/sol005/VnfPkgTest.java
@@ -30,9 +30,9 @@ import com.ubiqube.etsi.mano.config.Http403EntryPoint;
 import com.ubiqube.etsi.mano.controller.vnf.VnfPackageManagement;
 import com.ubiqube.etsi.mano.model.vnf.sol005.VnfPkgInfo;
 import com.ubiqube.etsi.mano.repository.VnfPackageRepository;
-import com.ubiqube.etsi.mano.service.EventManager;
 import com.ubiqube.etsi.mano.service.ManufacturerModel;
 import com.ubiqube.etsi.mano.service.Patcher;
+import com.ubiqube.etsi.mano.service.event.EventManager;
 
 @AutoConfigureMockMvc
 @WebMvcTest
